+++ src/app/basicauction/basicauction.tsx
@@ -1,17 +1,10 @@
 'use client';
 
-import { useEffect } from 'react';
-
 import { useSearchParams } from 'next/navigation';
 
 import useGetBasicAuctionList from '@/apis/queryHooks/basicAuction/useGetBasicAuctionList';
-import AuctionDropDown from '@/app/basicauction/components/auctiondropdown';
-import SearchBar from '@/app/basicauction/components/searchbar';
-import CheckIcon from '@/assets/svg/check.svg';
 import AuctionCard from '@/components/AuctionCard';
 import ListLayout from '@/components/ListLayout';
-import useBooleanState from '@/hooks/useBooleanState';
-import useSetSearchParam from '@/hooks/useSetSearchParam';
 import { AUCTIONPARAM_KEY } from '@/static/queryParam';
 
 import * as S from './Basicauction.css';
@@ -19,17 +12,6 @@ import * as S from './Basicauction.css';
 function BasicAuction() {
   const searchParams = useSearchParams();
   const searchKeywordParam = searchParams.get(AUCTIONPARAM_KEY.Q);
-
-  const { value: isChecked, toggle } = useBooleanState(false);
-  const { setSingleSearchParam } = useSetSearchParam();
-
-  useEffect(() => {
-    if (isChecked) {
-      setSingleSearchParam(AUCTIONPARAM_KEY.AUCTIONSTATUS, 'BIDDING');
-    } else {
-      setSingleSearchParam(AUCTIONPARAM_KEY.AUCTIONSTATUS, '');
-    }
-  }, [isChecked]);
 
   const { data } = useGetBasicAuctionList({
     title: searchKeywordParam || '',
@@ -44,36 +26,27 @@ function BasicAuction() {
 
   return (
     <div className={S.container}>
-      <section className={S.leftSection}>
+      <div className={S.searchContainer}>
         <div className={S.count}>
           <span>전체</span>
           <span>{data.result_data.content.length}</span>
         </div>
-        <SearchBar />
-        <label htmlFor="checkbox" className={`${S.label} ${isChecked ? S.checked : S.nonChecked}`}>
-          경매중
-          <CheckIcon />
-          <input id="checkbox" type="checkbox" className={S.checkbox} onChange={toggle} />
-        </label>
-        <AuctionDropDown />
-      </section>
-      <section className={S.rightSection}>
-        <ListLayout>
-          {data.result_data.content.map(item => (
-            <AuctionCard
-              key={item.auction_id}
-              id={item.auction_id}
-              image={item.image_keys}
-              title={item.title}
-              isLike={false}
-              startPrice={item.start_price}
-              startTime={item.start_date}
-              endTime={item.end_date}
-              nowPrice={item.now_price}
-            />
-          ))}
-        </ListLayout>
-      </section>
+      </div>
+      <ListLayout>
+        {data.result_data.content.map(item => (
+          <AuctionCard
+            key={item.auction_id}
+            id={item.auction_id}
+            image={item.image_keys}
+            title={item.title}
+            isLike={false}
+            startPrice={item.start_price}
+            startTime={item.start_date}
+            endTime={item.end_date}
+            nowPrice={item.now_price}
+          />
+        ))}
+      </ListLayout>
     </div>
   );
 }
