+++ src/app/basicauction/components/searchbar.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useRouter, useSearchParams } from 'next/navigation';
 
 import DeleteIcon from '@/assets/svg/delete.svg';
